+++ src/app/(main)/files/[id]/page.tsx
@@ -1,13 +1,13 @@
 import { notFound } from 'next/navigation';
 import { Flexbox } from 'react-layout-kit';
 
-import FileDetail from '@/app/(main)/files/features/FileDetail';
 import FileViewer from '@/features/FileViewer';
 import { createCallerFactory } from '@/libs/trpc';
 import { lambdaRouter } from '@/server/routers/lambda';
 import { PagePropsWithId } from '@/types/next';
 import { getUserAuth } from '@/utils/server/auth';
 
+import FileDetail from '../features/FileDetail';
 import Header from './Header';
 
 const createCaller = createCallerFactory(lambdaRouter);
